Added auto insert Apache settings
Added code to detect on Apache if the .htaccess is writable. If it is then auto insert the needed Apache rules. If not, notify the user.

Notice: FOR NGINX USERS
I'm not sure how to detect if the hosting env is Nginx. So if the if the .htaccess isn't writeable the plugin will always self deactivate itself. This means currently with Nginx users the plugin will fail everytime. 

Maybe you can adjust it to work for both envs out of the box?